Break down network module & move tmdb logic to a separate module.
