+++ shared/core/network/src/androidMain/kotlin/com/thomaskioko/tvmaniac/network/di/Module.kt
@@ -1,0 +1,6 @@
+package com.thomaskioko.tvmaniac.network.di
+
+import org.koin.core.module.Module
+import org.koin.dsl.module
+
+actual fun networkPlatformModule(): Module = module { }
