+++ shared/core/network/src/commonMain/kotlin/com/thomaskioko/tvmaniac/network/KtorClientFactory.kt
@@ -1,0 +1,7 @@
+package com.thomaskioko.tvmaniac.network
+
+import io.ktor.client.HttpClient
+
+expect class KtorClientFactory() {
+    fun build(httpUrl: String): HttpClient
+}
