+++ shared/core/network/src/commonMain/kotlin/com/thomaskioko/tvmaniac/network/di/Module.kt
@@ -1,0 +1,5 @@
+package com.thomaskioko.tvmaniac.network.di
+
+import org.koin.core.module.Module
+
+expect fun networkPlatformModule(): Module
