--- shared/core/remote/src/androidMain/kotlin/com/thomaskioko/tvmaniac/remote/di/Module.kt
@@ -1,6 +1,0 @@
-package com.thomaskioko.tvmaniac.remote.di
-
-import org.koin.core.module.Module
-import org.koin.dsl.module
-
-actual fun remotePlatformModule(): Module = module { }
