--- shared/core/remote/src/commonMain/kotlin/com/thomaskioko/tvmaniac/remote/KtorClientFactory.kt
@@ -1,7 +1,0 @@
-package com.thomaskioko.tvmaniac.remote
-
-import io.ktor.client.HttpClient
-
-expect class KtorClientFactory() {
-    fun build(): HttpClient
-}
