--- shared/core/remote/src/commonMain/kotlin/com/thomaskioko/tvmaniac/remote/di/Module.kt
@@ -1,5 +1,0 @@
-package com.thomaskioko.tvmaniac.remote.di
-
-import org.koin.core.module.Module
-
-expect fun remotePlatformModule(): Module
