--- shared/core/remote/src/iosMain/kotlin/com/thomaskioko/tvmaniac/remote/di/Module.kt
@@ -1,9 +1,0 @@
-package com.thomaskioko.tvmaniac.remote.di
-
-import com.thomaskioko.tvmaniac.remote.KtorClientFactory
-import org.koin.core.module.Module
-import org.koin.dsl.module
-
-actual fun remotePlatformModule(): Module = module {
-    single { KtorClientFactory().build() }
-}
