+++ shared/domain/tmdb/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/tmdb/api/model/ShowDetailResponse.kt
